Fixes bug in exchange and loans lists.
Exchanges from institutions without a code become inaccessable, since there is no code to click on to expand the list of gifts/exchanges from that institution. Loans from institutions without a code are not inaccessible, but there is no way to tell where they are from without editing the.

This fixes both problems by preferring the code to show an institution, but falling back to the institution name if code is not specified.