Get institution data from GBIF GrSciColl API
This adds a button to the Institution editor to pull in data for an institution code from GBIF's GrSciColl API, which includes all institutions in Index Herbariorum.

In addition to adding data to all of the institution fields in symbiota, it adds additional information into the notes field that would be useful for collection managers: incorporated collections, taxonomic and geographic focus, number of specimens, and links to the collection page on GrSciColl and Index Herbariorum (if present).

In some cases there are multiple results (collections, institutions, different data sources) for an institution code. This allows the user to pick which result they would like to use (thanks Frank Bungartz). .