Adds ability to edit loan/exchange shipping invoices
Adds an edit button to the shipping invoice form. Clicking it allows a user to edit the text, use bold and italics, paste in rich text or images, etc. The edited invoice can then be printed, but edits are not saved. Mostly, this is useful to add additional text to the shipping invoice beyond what is possible to auto-generate.

It uses the same nice solution that @arbolitoloco built for the label maker for editing labels before printing.